requirements.txt
# Core Python packages for IDS system
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
psutil>=5.8.0
kafka-python>=2.0.2
requests>=2.26.0
pyyaml>=6.0
click>=8.0.0

# Network analysis
scapy>=2.4.5
pyshark>=0.4.5

# Machine Learning
tensorflow>=2.8.0
torch>=1.10.0

# Monitoring and logging
prometheus-client>=0.12.0
structlog>=21.5.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.12.0
flake8>=4.0.0
mypy>=0.930
